Create somethinguseful
This is an attempt to use the fullstory rest api along with the full story sdk here on git hub.  This is my attempt at using logic and not any actual coding to use the rest api output.  In reality this won't work but as stated is an attempt for me to understand both fullstory and github and what I can possibly do given time.